requirements.txt
Django
django-cors-headers
django-clearcache
djangorestframework
whitenoise
environ
django-environ
python-dotenv
cloudinary
django-rest-swagger
Django-Cloudinary-storage
django-storages
Pillow
drf-yasg